Move Authentication methods to class "Authentication"
