+++ src/App.php
@@ -8,6 +8,7 @@ use Exception;
 use Friendica\App\Arguments;
 use Friendica\App\BaseURL;
 use Friendica\App\Page;
+use Friendica\Core\Authentication;
 use Friendica\Core\Config\Cache\ConfigCache;
 use Friendica\Core\Config\Configuration;
 use Friendica\Core\Config\PConfiguration;
@@ -717,7 +718,7 @@ class App
 				Model\Profile::openWebAuthInit($token);
 			}
 
-			Login::sessionAuth();
+			Authentication::sessionAuth();
 
 			if (empty($_SESSION['authenticated'])) {
 				header('X-Account-Management-Status: none');
